+++ sphinxcontrib/sphinx_bazel/domain/workspace.py
@@ -14,23 +14,24 @@ class BazelWorkspace(Directive):
     """
     Directive to mark description of a new workspace.
     """
-    
+
     has_content = True
     required_arguments = 1
     optional_arguments = 0
     final_argument_whitespace = True
     option_spec = {
-        'path':  directives.unchanged,
-        'hide': directives.flag
+        'path': directives.unchanged,
+        'hide': directives.flag,
+        'show_type': directives.flag
     }
-    
+
     def run(self):
         env = self.state.document.settings.env
         workspace_name = self.arguments[0].strip()
         workspace_path = self.options.get('path', '')
         env.ref_context['bazel:workspace'] = workspace_name
         ret = []
-        
+
         env.domaindata['bazel']['workspaces'][workspace_name] = (env.docname,
                                                                  workspace_path)
         # make a duplicate entry in 'objects' to facilitate searching for
@@ -50,13 +51,20 @@ class BazelWorkspace(Directive):
         if self.options.get('hide', False) is None:
             # No output is wanted
             return ret
-            
+
         workspace_string = workspace_name
+        if self.options.get('show_type', False) is None:
+            sig_type_string = 'workspace: '
+            ret.append(addnodes.desc_name(sig_type_string, sig_type_string,
+                                          classes=['bazel', 'type', 'workspace']))
+
         if workspace_path:
             workspace_string += ' ({})'.format(workspace_path)
+
         workspace_name_node = addnodes.desc_name(workspace_string, workspace_string)
+
         ret.append(workspace_name_node)
-        
+
         contentnode = addnodes.desc_content()
         ret.append(contentnode)
         self.state.nested_parse(self.content, self.content_offset, contentnode)